Move Harvest's Drush and Helper into namespace Drupal\harvest\Commands
